+++ FieldtypeSnipWireTaxSelector/FieldtypeSnipWireTaxSelector.module.php
@@ -17,7 +17,7 @@ namespace ProcessWire;
  *
  */
 
-require_once dirname(__DIR__) . '/helpers/Taxes.php';
+wire('classLoader')->addNamespace('SnipWire\Helpers', dirname(__DIR__) . '/helpers');
 
 use SnipWire\Helpers\Taxes;